+++ src/ocamlorg/lib/ocamlorg.mli
@@ -3,6 +3,6 @@
 module Opam_user = Opam_user
 module Package = Package
 
-val topelevels_path : Fpath.t
+val toplevels_path : Fpath.t
 (** The path where the toplevel scripts are located. Delete when they are served
     from a CDN. *)