+++ src/ocamlorg_toplevel/bin/toplevel_stdlib.ml
@@ -1,0 +1,10 @@
+open Brr
+
+let () =
+  let button =
+    Document.find_el_by_id G.document (Jstr.v "toplevel-load") |> Option.get
+  in
+  Ev.listen
+    Ev.click
+    (fun _ -> Ocamlorg_toplevel.Toplevel.run "/toplevels/worker_base.js")
+    (El.as_target button)