+++ src/ocamlorg_toplevel/bin/worker_stdlib.ml
@@ -1,0 +1,1 @@
+let () = Worker.run ()